+++ search-films/src/context/user.context/index.tsx
@@ -1,5 +1,5 @@
+import UserProvider from '../provider/user.provider/user.provider';
 import UserContext from './user.context';
-import UserProvider from './user.provider';
 export type { IUserContext } from './user.context.d';
 export { UserContext, UserProvider };