+++ search-films/src/context/user.context/user.context.d.ts
@@ -1,6 +1,6 @@
 export interface IUserContext {
     currentUser: string | null;
     isAuthenticated: boolean;
-    // TODO: ?
+    // TODO: (me) возвращаемый тип?
     logoutCurrentUser?: () => void;
 };